feat(types): [STABLE] ban all same-width mixed signed/unsigned arithmetic (closes #70)
Previously only 64-bit same-width pairs (i64+u64) and cross-width
i64+u32 were banned. Now ALL same-width mixed pairs are banned for
let variables: i8+u8, i16+u16, i32+u32, i64+u64.

Rationale: in TS there's no unsigned — number+number=number. C's
unsigned-wins rule silently converts -1 to 4294967295, violating
TS expectations (П2) and C safety (П3). Now requires explicit 'as'.

const/literals remain exempt (compile-time value analysis).

TDD: 3 new error tests, 1 converted from ok to error, 2 new positive
tests (const exempt, explicit as). Spec updated in same commit.

+++ test/cases/phase2/widening/ok-explicit-as-i32-u32/expected.c
@@ -4,7 +4,7 @@ int main(void) {
     TSC_INIT();
     int32_t a = 1;
     uint32_t b = 2U;
-    uint32_t x = a + b;
-    printf("%u\n", x);
+    int32_t x = (int32_t)((uint32_t)a + (uint32_t)(int32_t)b);
+    printf("%d\n", x);
     return 0;
 }
